README: Improved readability and improved syntax
Didn't change any content! Simply improved some of the syntax and also added some markup to make certain keywords standout so that it's easier to read. Improved some of the formatting and whatnot. Awesome repo! I know this hasn't been touched in a while, but I'm very intrigued by it!